+++ pywaffle/__init__.py
@@ -2,6 +2,7 @@
 # -*-coding: utf-8 -*-
 
 from ._version import __version__
+from .functional import waffle
 from .waffle import Waffle
 
-__all__ = ["Waffle", "__version__"]
+__all__ = ["Waffle", "waffle", "__version__"]